finish chapter 20-file-upload-and-download

+++ controllers/admin.js
@@ -1,6 +1,7 @@
-const Product = require('../models/products')
-const { validationResult } = require('express-validator')
 const mongoose = require('mongoose')
+const { validationResult } = require('express-validator')
+const Product = require('../models/products')
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -15,7 +16,19 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const userId = req.user
-  const { title, price, image, description } = req.body
+  const { title, price, description } = req.body
+  const image = req.file
+  // console.log(image)
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      product: { title, price, description },
+      validationErrors: [],
+      errorMessage: 'Attached file is not a valid image'
+    })
+  }
   const errors = validationResult(req)
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -27,10 +40,11 @@ exports.postAddProduct = (req, res, next) => {
       errorMessage: null
     })
   }
+  const imageUrl = image.filename
   const product = new Product({
     title: title,
     price: price,
-    image: image,
+    imageUrl: imageUrl,
     description: description,
     userId: userId // or use userId: req.user, mongoose will do the rest
   })
@@ -111,18 +125,20 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const userId = req.user
   const productId = req.body.productId
-  const { title, price, imageUrl, description } = req.body
+  const { title, price, description } = req.body
+  const image = req.file
   const errors = validationResult(req)
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
       path: '/admin/edit-product',
       editing: true,
-      product: { _id: productId, title, price, imageUrl, description },
+      product: { _id: productId, title, price, description },
       validationErrors: errors.array(),
       errorMessage: null
     })
   }
+
   Product.findById(productId)
     .then(product => {
       if (!product.userId.equals(req.user._id)) {
@@ -130,9 +146,12 @@ exports.postEditProduct = (req, res, next) => {
       }
       product.title = title
       product.price = price
-      product.imageUrl = imageUrl
       product.description = description
       product.userId = userId
+      if (image) {
+        fileHelper.deleteFile(`images/${product.imageUrl}`)
+        product.imageUrl = image.filename
+      }
       return product.save()
         .then(result => {
           console.log('Product updated successfully')
@@ -147,13 +166,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId
-  Product.deleteOne({ _id: productId, userId: req.user._id })
-    .then(() => {
-      res.redirect('/admin/products')
-    })
-    .catch(err => {
-      if (err) {
-        console.error(err)
-      }
-    })
+  Product.findById(productId)
+  .then(product => {
+    if (!product) {
+      return next(new Error('Product not found'))
+    }
+    fileHelper.deleteFile(`images/${product.imageUrl}`)
+    return Product.deleteOne({ _id: productId, userId: req.user._id })
+  })
+  .then(() => {
+    console.log('Product deleted successfully')
+    res.redirect('/admin/products')
+  })
+  .catch(err => {
+    const error = new Error(err)
+    error.httpStatusCode = 500
+    return next(error)
+  })
 }
